Cleanup

+++ fmt_strings.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <string.h>
 
 int main() {
    char string[10];
@@ -11,8 +12,8 @@ int main() {
    printf("[A] Dec: %d, Hex: %x, Unsigned: %u\n", A, A, A);
    printf("[B] Dec: %d, Hex: %x, Unsigned: %u\n", B, B, B);
    printf("[field width on B] 3: '%3u', 10: '%10u', '%08u'\n", B, B, B);
-   printf("[string] %s  Address %08x\n", string, string);
+   printf("[string] %s  Address %08x\n", string, (unsigned int)(string));
 
    // Example of unary address operator (dereferencing) and a %x format string
-   printf("variable A is at address: %08x\n", &A);
+   printf("variable A is at address: %08x\n", (unsigned int)&A);
 }
