+++ pointer_types5.c
@@ -6,9 +6,9 @@ int main() {
 	char char_array[5] = {'a', 'b', 'c', 'd', 'e'};
 	int int_array[5] = {1, 2, 3, 4, 5};
 
-	unsigned int hacky_nonpointer;
+	unsigned long hacky_nonpointer;
 
-	hacky_nonpointer = (unsigned int) char_array;
+	hacky_nonpointer = (unsigned long) char_array;
 
 	for(i=0; i < 5; i++) { // iterate through the int array with the int_pointer
 		printf("[hacky_nonpointer] points to %p, which contains the char '%c'\n",
@@ -16,8 +16,8 @@ int main() {
 		hacky_nonpointer = hacky_nonpointer + sizeof(char);
 	}
 
-	hacky_nonpointer = (unsigned int) int_array;
-	
+	hacky_nonpointer = (unsigned long) int_array;
+
 	for(i=0; i < 5; i++) { // iterate through the int array with the int_pointer
 		printf("[hacky_nonpointer] points to %p, which contains the integer %d\n",
             hacky_nonpointer, *((int *) hacky_nonpointer));
